wip: Timeline scroll behaviour
