+++ packages/apps/spaces/components/shared/location/location-list.module.scss
@@ -17,4 +17,5 @@
   align-items: center;
   flex: 1;
   margin-bottom: 8px;
+  font-size: var(--font-size-sm);
 }
