+++ packages/apps/spaces/components/ui-kit/organisms/timeline/context/useTimeline.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { TimelineContext } from './timelineContext';
+
+export const useTimeline = () => useContext(TimelineContext);
